Dev: sbd: Split class SBDManager into sbd.py from bootstrap.py
